Add sys-apps/util-linux overlay
The util-linux ebuild fails to cross compile properly because the
HAVE_SCANF_AS_MODIFIER define is missing. Because of that the UL_SCNsA
define does not exist which is used in two scanf invocations. This
change provides an ebuild that sets the scanf_cv_alloc_modifier variable
that is read by the configure infrastructure based on which the defines
will be set.