Update app-misc/tmux ebuild to 2.2
With version 2.2 the forkpty detection fails. The forkpty() function is
provided by libutil but the configure run is buggy and appends -lresolv
to LIBS before the forkpty check and libresolv does not exist. So the
check for forkpty fails because libresolv cannot be found and not
because libutil does not exist or does not provide forkpty().
This change takes a brute-force approach and removes the faulty LIBS
setting. Note that for some reason this removal has to happen after the
configure run occurred (the following make will trigger a new configure
and that will check for libutil successfully).